+++ addOns/zest/src/main/java/org/zaproxy/zap/extension/zest/dialogs/ZestParameterizeDialog.java
@@ -68,7 +68,7 @@ public class ZestParameterizeDialog extends StandardFieldsDialog {
 
     @Override
     public void save() {
-        this.extension.perameterize(
+        this.extension.parametrize(
                 script,
                 node,
                 request,