+++ addOns/zest/src/main/java/org/zaproxy/zap/extension/zest/menu/ZestPopupCommentOnOff.java
@@ -91,13 +91,13 @@ public class ZestPopupCommentOnOff extends ExtensionPopupMenuItem {
                         this.setEnabled(false);
                         return false;
                     } else if ((ZestZapUtils.getElement(node) instanceof ZestScript)) {
-                        // Cant comment the whole script
+                        // Can't comment the whole script
                         this.setEnabled(false);
                         return false;
                     } else if (ZestZapUtils.getShadowLevel(node) > 0) {
                         // Ignore these
                     } else if (!(ZestZapUtils.getElement(node) instanceof ZestStatement)) {
-                        // Cant comment these
+                        // Can't comment these
                         this.setEnabled(false);
                         break;
                     } else {